Création de l'interface classe Address et AddressImpl dans "com.cesi.bhs.api.db"
